fix: remove duplicate animator merger and add parameter type mismatch error asset references
Closes: #549
