--- Editor/AnimatorMerger.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 3c76e0db714645f7aa6580f208f98e49
-timeCreated: 1661644852
